first part of composing toot
I got a nice little chord progression thanks to the tutorial. This
part ends around https://youtu.be/lGs7JOOVjag?t=1468

It was interesting to work on this partially because debugging why a
sound wasn't working wasn't terribly easy. One of the things that
frustrates me a bit is that there's not seemingly an easy way to have
named arguments in the methods, which makes it kind of annoying to look
back at the code and remember what's going on.